refactor: Names 일급 컬렉션 생성을 통한이름 중복 검사 로직 분리
